ДЗ 2: добавление меню в остальные активити
