+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
         package="com.example.baforecast">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
             android:allowBackup="true"
@@ -21,8 +21,14 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SelectCityActivity"/>
-        <activity android:name=".SettingsActivity"/>
+        <activity
+                android:name=".SelectCityActivity"
+                android:label="@string/select_city_name"
+                android:theme="@style/Theme.BaForecast.NoActionBar" />
+        <activity
+                android:name=".SettingsActivity"
+                android:label="@string/settings_name"
+                android:theme="@style/Theme.BaForecast.NoActionBar" />
     </application>
 
 </manifest>
