+++ app/src/main/java/com/example/baforecast/BaseActivity.java
@@ -1,13 +1,26 @@
 package com.example.baforecast;
 
+import android.annotation.SuppressLint;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
 
-public class BaseActivity extends AppCompatActivity {
+import com.google.android.material.navigation.NavigationView;
+
+public class BaseActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
     private static final String NameSharedPreference = "LOGIN";
     private static final String IsDarkTheme = "IS_DARK_THEME";
+    private City city;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -16,6 +29,13 @@ public class BaseActivity extends AppCompatActivity {
             setTheme(R.style.AppDarkTheme);
         }else{
             setTheme(R.style.Theme_BaForecast_NoActionBar);
+        }
+
+        Intent intent = getIntent();
+        if (intent.hasExtra("parcel")) {
+            city = (City) intent.getSerializableExtra("parcel");
+        } else {
+            city = new City(getResources().getStringArray(R.array.cities)[0]);
         }
         setContentView(R.layout.activity_base);
     }
@@ -30,6 +50,90 @@ public class BaseActivity extends AppCompatActivity {
         SharedPreferences.Editor editor = sharedPref.edit();
         editor.putBoolean(IsDarkTheme, isDarkTheme);
         editor.apply();
+    }
+
+    protected Toolbar initToolbar() {
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        return toolbar;
+    }
+
+
+    protected void initDrawer(Toolbar toolbar) {
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        NavigationView navigationView = findViewById(R.id.nav_view);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+            this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.addDrawerListener(toggle);
+        toggle.syncState();
+        navigationView.setNavigationItemSelectedListener(this);
+    }
+
+
+    @SuppressLint("NonConstantResourceId")
+    @Override
+    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+        int id = item.getItemId();
+
+        switch (id) {
+            case R.id.nav_home:
+                Intent intentMain = new Intent(getApplicationContext(), MainActivity.class);
+                intentMain.putExtra(Constants.EXTRA_PARCEL, city);
+                startActivity(intentMain);
+                break;
+            case R.id.nav_select_city:
+                Intent intentSelectAct = new Intent(getApplicationContext(), SelectCityActivity.class);
+                intentSelectAct.putExtra(Constants.EXTRA_PARCEL, city);
+                startActivity(intentSelectAct);
+                break;
+            case R.id.nav_about:
+                Toast toast = new Toast(getApplicationContext());
+                toast.setText("J hfphf");
+                toast.show();
+                break;
+        }
+
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+        return true;
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.main, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+
+        if (id == R.id.action_settings) {
+            Intent intent = new Intent(getApplicationContext(), SettingsActivity.class);
+            startActivityForResult(intent, Constants.SETTING_CODE);
+            return true;
+        }
+
+        if (id == R.id.action_select_city) {
+            Intent intent = new Intent(getApplicationContext(), SelectCityActivity.class);
+            intent.putExtra(Constants.EXTRA_PARCEL, city);
+            startActivity(intent);
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public void onBackPressed() {
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+
+        if (drawer.isDrawerOpen(GravityCompat.START)) {
+            drawer.closeDrawer(GravityCompat.START);
+        } else {
+            super.onBackPressed();
+        }
+
     }
 
 }
