+++ app/src/main/java/com/example/baforecast/Constants.java
@@ -3,4 +3,5 @@ package com.example.baforecast;
 public class Constants {
     public static final String EXTRA_PARCEL = "parcel";
     public static final String EXTRA_CURRENT_CITY = "CurrentCity";
+    public static final int SETTING_CODE = 88;
 }
