+++ app/src/main/java/com/example/baforecast/MainActivity.java
@@ -6,32 +6,16 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.util.Log;
-import android.view.ContextMenu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
-import android.view.Menu;
-import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.example.baforecast.model.WeatherRequest;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.navigation.NavigationView;
 import com.google.gson.Gson;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.appcompat.app.AlertDialog;
-import androidx.core.view.GravityCompat;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
-import androidx.navigation.ui.NavigationUI;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -55,7 +39,6 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
     private static final int DAY_COUNT = 5;
     private static final String WEATHER_URL = "https://api.openweathermap.org/data/2.5/weather?q=%s,RU&appid=%s";
     private City city;
-    private static final int SETTING_CODE = 88;
     private TextView txtViewCity;
     private TextView txtViewTemperature;
     private TextView txtViewPressure;
@@ -68,9 +51,9 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
         Toolbar toolbar = initToolbar();
         initDrawer(toolbar);
 
-        txtViewCity = (TextView) findViewById(R.id.text_city);
-        txtViewTemperature = (TextView) findViewById(R.id.text_curr_temp);
-        txtViewPressure = (TextView) findViewById(R.id.text_pressure);
+        txtViewCity = findViewById(R.id.text_city);
+        txtViewTemperature = findViewById(R.id.text_curr_temp);
+        txtViewPressure = findViewById(R.id.text_pressure);
 
         Intent intent = getIntent();
         if (intent.hasExtra("parcel")) {
@@ -87,49 +70,6 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
 
         initRecyclerView(generateDays(), generateTemperatures());
 
-    }
-
-    private Toolbar initToolbar() {
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        return toolbar;
-    }
-
-
-    private void initDrawer(Toolbar toolbar) {
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-            this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
-        toggle.syncState();
-        navigationView.setNavigationItemSelectedListener(this);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.main, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        int id = item.getItemId();
-
-        if (id == R.id.action_settings) {
-            Intent intent = new Intent(MainActivity.this, SettingsActivity.class);
-            startActivityForResult(intent, SETTING_CODE);
-            return true;
-        }
-
-        if (id == R.id.action_select_city) {
-            Intent intent = new Intent(getApplicationContext(), SelectCityActivity.class);
-            intent.putExtra(Constants.EXTRA_PARCEL, city);
-            startActivity(intent);
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
     }
 
     private void initRecyclerView(String[] days, String[] temperatures) {
@@ -169,7 +109,7 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == SETTING_CODE) {
+        if (requestCode == Constants.SETTING_CODE) {
             recreate();
         }
     }
@@ -252,37 +192,4 @@ public class MainActivity extends BaseActivity implements NavigationView.OnNavig
         alert.show();
     }
 
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        int id = item.getItemId();
-
-        switch (id) {
-            case R.id.nav_select_city:
-                Intent intent = new Intent(getApplicationContext(), SelectCityActivity.class);
-                intent.putExtra(Constants.EXTRA_PARCEL, city);
-                startActivity(intent);
-                break;
-            case R.id.nav_about:
-                Toast toast = new Toast(getApplicationContext());
-                toast.setText("J hfphf");
-                toast.show();
-                break;
-        }
-
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-
-    }
 }
