+++ app/src/main/java/com/example/baforecast/SelectCityActivity.java
@@ -3,12 +3,14 @@ package com.example.baforecast;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 
 import com.example.baforecast.observer.Observer;
@@ -35,6 +37,8 @@ public class SelectCityActivity extends BaseActivity implements View.OnClickList
 
         publisher.subscribe(this);
         setContentView(R.layout.activity_select_city);
+        Toolbar toolbar = initToolbar();
+        initDrawer(toolbar);
 
         Intent intent = getIntent();
 
@@ -55,6 +59,26 @@ public class SelectCityActivity extends BaseActivity implements View.OnClickList
 
         Button selectCity = (Button) findViewById(R.id.buttonSelectCity);
         selectCity.setOnClickListener(this);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        int id = item.getItemId();
+
+        if (id == R.id.action_settings) {
+            Intent intent = new Intent(getApplicationContext(), SettingsActivity.class);
+            startActivityForResult(intent, Constants.SETTING_CODE);
+            return true;
+        }
+
+        if (id == R.id.action_select_city) {
+            Intent intent = new Intent(getApplicationContext(), SelectCityActivity.class);
+            intent.putExtra(Constants.EXTRA_PARCEL, city);
+            startActivity(intent);
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
