+++ app/src/main/java/com/example/baforecast/SettingsActivity.java
@@ -8,6 +8,7 @@ import android.widget.Button;
 import android.widget.Toast;
 
 import androidx.appcompat.widget.SwitchCompat;
+import androidx.appcompat.widget.Toolbar;
 
 import com.google.android.material.textfield.TextInputEditText;
 
@@ -17,6 +18,8 @@ public class SettingsActivity extends BaseActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
+        Toolbar toolbar = initToolbar();
+        initDrawer(toolbar);
 
         SwitchCompat switchDarkTheme = findViewById(R.id.switch1);
         switchDarkTheme.setChecked(isDarkTheme());
