+++ app/src/main/res/layout/activity_select_city.xml
@@ -1,64 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
+        android:id="@+id/drawer_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SelectCityActivity">
+        android:fitsSystemWindows="true"
+        tools:openDrawer="start">
 
-    <ScrollView
-            android:id="@+id/scroll"
-            android:layout_width="fill_parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintBottom_toTopOf="@+id/linear"
-            android:layout_height="0px">
-
-        <FrameLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-            <fragment
-                    android:name="com.example.baforecast.CitiesFragment"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:id="@+id/cities" />
-        </FrameLayout>
-    </ScrollView>
-
-    <LinearLayout
-            android:id="@+id/linear"
+    <include
+            layout="@layout/app_bar_select_city"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintBottom_toTopOf="@id/buttonSelectCity">
+            android:layout_height="match_parent" />
 
-        <CheckBox
-                android:id="@+id/checkBoxPressure"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/leftIndent"
-                android:layout_weight="1"
-                android:text="@string/label_pressure"
-                android:theme="@style/AdditionText" />
+    <com.google.android.material.navigation.NavigationView
+            android:id="@+id/nav_view"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_gravity="start"
+            android:fitsSystemWindows="true"
+            app:headerLayout="@layout/nav_header_main"
+            app:menu="@menu/activity_main_drawer" />
 
-        <CheckBox
-                android:id="@+id/checkBoxWindSpeed"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="@string/label_wind_speed"
-                android:theme="@style/AdditionText" />
-
-    </LinearLayout>
-
-    <com.google.android.material.button.MaterialButton
-            android:id="@+id/buttonSelectCity"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/button_select"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</androidx.drawerlayout.widget.DrawerLayout>
