+++ app/src/main/res/values-ru/strings.xml
@@ -50,4 +50,7 @@
     </string-array>
     <string name="alert_title">Внимание!</string>
     <string name="alert_button">OK</string>
+    <string name="select_city_name">Выберите город</string>
+    <string name="settings_name">Настройки</string>
+    <string name="button_go_to_home">Домой</string>
 </resources>
