+++ app/src/main/res/values/strings.xml
@@ -26,6 +26,9 @@
     <string name="back">Back</string>
     <string name="alert_title">WARNING!</string>
     <string name="alert_button">OK</string>
+    <string name="select_city_name">Select city</string>
+    <string name="settings_name">Settings</string>
+    <string name="button_go_to_home">Home</string>
     <string-array name="cities">
         <item>Moscow</item>
         <item>Saint Petersburg</item>
